Add ControllerSpecificationGenerator
Add ControllerSpecificationGenerator to the container and console command.
The spec generated by describe:controller were extending ObjectBehaviour.
This meant that the end user needed to add boilerplate code to each spec
to inject stub Request and Response. This is now done in the ControllerBehviour
class that all Controller specs will now extend.